pyproject.toml
[project]
name = "wabi-agent-backend-new"
version = "2.1.0"
description = "Gemini-powered LangGraph Orchestrator"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.60",
    "langgraph-checkpoint>=2.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.3",
    "openai>=1.59.6",
    "boto3>=1.35.90",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.30.0",
    "pillow>=10.0.0",
    "protobuf >= 3.20.2, < 6"
]

[project.optional-dependencies]
dev = [
    "langgraph-cli>=0.1.55",
    "langgraph-api>=0.7.13"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph_app"]
